+++ packages/dartway_serverpod_core/dartway_serverpod_core_flutter/lib/core/dw_core.dart
@@ -43,6 +43,7 @@ class DwCore<
     required this.client,
     required this.dwAlerts,
     required this.getUserId,
+    super.plugins,
     this.onStreamingStatusChanged,
   }) {
     setDwInstance(this);